requirements.txt
fastapi
uvicorn
requests
faiss-cpu
numpy
torch
transformers
pdfplumber
python-docx